fix : 버스트 순서에 따른 승패 결과 수정
